feat(spdk): wire v2 env-tunables into the data path
Restore the LONGHORN_V2_* tunable apparatus and its apply sites that were
dropped during the upstream re-port:

- types.go: const->var for the replica/rebuild timeouts plus new tunables
  (iobuf pool counts, transport TOS, accel_mlx5 mkeys-per-core, lvol
  clear_method / thin_provision, lvstore cluster size), with init()
  reading the LONGHORN_V2_* env overrides and the accelMlx5NumRequests()
  / spdkCoreCount() helpers.
- replica.go/backing_image.go: pass defaultLvolClearMethod +
  defaultThinProvision to the bdev_lvol_create sites instead of hardcoded
  "", true.
- disk.go: create lvstores with defaultLvstoreClusterSize.
- replica.go: the cloning, rebuilding-src and external-snapshot remote
  attaches now use connectNVMfBdevWithReconnect with r.transport() and
  the rebuild-path timeouts, restoring RDMA (the bare connectNVMfBdev was
  TCP-hardcoded) and sub-second teardown-during-rebuild failover.

replicaCtrlrLossTimeoutSec stays at the tuned 3s (down from upstream 15s)
to avoid reactor saturation when a remote replica IM disappears.

server.go startup tuning (iobuf/TOS/mlx5/framework_start_init) is a
separate change coupled to the IM --wait-for-rpc entrypoint flag.

+++ pkg/spdk/util.go
@@ -162,7 +162,7 @@ func connectNVMfBdev(spdkClient *spdkclient.Client, controllerName, address stri
 				spdktypes.NvmeTransportTypeTCP,
 				adrfam,
 				int32(ctrlrLossTimeout),
-				replicaReconnectDelaySec,
+				int32(replicaReconnectDelaySec),
 				int32(fastIOFailTimeoutSec),
 				replicaMultipath,
 			)
